okkkkaayyy

+++ task1/best_metrics.py
@@ -53,10 +53,10 @@ for i in tqdm(range(0, num_categories), desc='Getting Metrics'):
     cm = confusion_matrix(class_test, class_preds)
 
     if len(cm) != 1:
-        tp = long(cm[0][0])
-        fp = long(cm[0][1])
-        tn = long(cm[1][1])
-        fn = long(cm[1][0])
+        tp = int(cm[0][0])
+        fp = int(cm[0][1])
+        tn = int(cm[1][1])
+        fn = int(cm[1][0])
 
         # Precision
         if tp == 0 and fp == 0:
